+++ src/App.tsx
@@ -1,3 +1,3 @@
-const App = () => <>Boilerplate Vite React TS</>;
+const App = () => <h1>create-boilerplate-vite-react-ts</h1>;
 
 export default App;