+++ src/pages/Contact.tsx
@@ -3,7 +3,8 @@ import ROUTES from "@/constants/routes";
 
 const Contact = () => (
   <>
-    <h1>Contact</h1>
+    <h1>create-boilerplate-vite-react-ts</h1>
+    <h2>Contact</h2>
     <Link to={ROUTES.home}>Go to home</Link>
   </>
 );