+++ src/pages/Home.tsx
@@ -3,7 +3,8 @@ import ROUTES from "@/constants/routes";
 
 const Home = () => (
   <>
-    <h1>Home</h1>
+    <h1>create-boilerplate-vite-react-ts</h1>
+    <h2>Home</h2>
     <Link to={ROUTES.contact}>Go to contact</Link>
   </>
 );