+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # [Versions](https://github.com/Tracktor/eslint-config-react-tracktor/releases)
 
 
-## v2.5.0
-- **[feat]** :let user choose package manager
+## v2.5.1
+- **[style]** : script create app
+- **[feat]** : add break line to log script when is running
+- **[feat]** : minor changes